TextUI.cs: Patch for Mono environments absent of nunit24, xunit (#24)
* src/**/*.csproj: Use AppendConstants within DefineConstants declations

This patch provides support for an AppendContants build parameter, in
NUnitLite under MSBuild. This allows for building with such as the
following build options, under Mono:

~~~~
msbuild src/framework/nunitlite-4.5.csproj \
  /p:AppendConstants=MONO \
  /restore /t:build /p:CscToolPath=${PREFIX}/lib/mono/4.5/ \
  /p:TargetFrameworkVersion=4.7.2 /p:SignAssembly=false /v:detailed
~~~~

To some extent, this may serve to emulate some of the build
configuration under NAnt.

* TextUI.cs: Patch for Mono environments absent of nunit24, xunit

This patch extends on a patch published in an earlier branch, available
via contrib under the branch ci/patch-append-constants in changeset 9d3568e

Concerning nunit24 support for Mono in NUnitLite, the nunit24 BabySitter
class was removed from the Mono mcs source tree, in Mono 6.6.0.161
https://github.com/mono/mono/commit/55fbd14763ba5e6ed6dc63326d39782ff76afbe9

Concering xunit support for Mono in NUnitLite, the mono xunit-binaries
external may not be available under some Mono installations.

This changeset applies a couple of additional preprocessor declarations
in NUnitLite TextUI.cs, such that may serve to allow for building in a
Mono environment without nunit24 support and/or without xunit support,
vis a vis
MONO_NO_NUNIT24
MONO_NO_XUNIT

This may be applied with such as the following msbuild cmd - noting the
particular syntax for providing a sequence of constant symbols in
AppendConstants, namely double quoting the value as delimited with comma
characters, such that may be supported in UNIX-like shell environments:

~~~~
msbuild src/framework/nunitlite-4.5.csproj /restore /t:build \
 /p:CscToolPath=/usr/pkg/lib/mono/4.5/ /p:SignAssembly=false /v:detailed \
 /p:RunCodeAnalysis=false /p:TargetFrameworkVersion=4.7.2 \
 /p:AppendConstants=\"MONO,MONO_NO_NUNIT24,MONO_NO_XUNIT\"
~~~~

This may serve to prevent a build failure, such that may otherwise occur
while compiling TextUI.cs when the MONO symbol is declared in the
compiler environment.

For Mono environments in which the xunit assemblies may be available,
of course the MONO_NO_XUNIT symbol may be removed from the declaration
of the AppendConstants property. The similar would apply, repectively,
for any Mono environment in which the original nunit24 support may be
available.

For purposes of ensuring backwards compatibility, by default, the build
under MONO would be produced as in the original behavior, i.e assuming
nunit24 and xunit would both be available in the build environment.